# Wrap backslash processing in a branch