This wraps `find_escaped` in `if (backslash || prev_escape)`, short-circuiting roughly 9 instructions per loop iteration. I did not add an `if` for quote processing, because it'll only skip the clmul and strings are ridiculously common (I feel like it'll be rare for 64-bytes to *ever* lack a string, except in cases of > 64 byte strings).

I know this is all about branch-free, but ... backslashes just aren't that common. And in files where they happen a reasonable amount, they are going to be on the edge of predictable: like, they'll happen basically everywhere (multiline strings that use \n's all over), in runs (small spurts of strings that have weird characters), or in repeated patterns (because they are all objects of the same type.

This should be a discussion; you know a *lot* more about the impact of branches than I do, but this was striking enough that I felt the need to bring it up.

Here are the numbers I got on this machine, for stage 1, which improves it by 5-10% across the board:

| File | master GB/s stage 1| `if(backslash)` GB/s stage 1 |
|---|---|---|
| apache_builds | 6.526923 | 6.993132 |
| canada | 5.698863 | 6.097105 |
| citm_catalog | 7.272438 | 7.883177 |
| github_events | 6.714639 | 7.236889 |
| gsoc-2018 | 7.793515 | 8.407860 |
| instruments | 6.461760 | 6.779877 |
| marine_ik | 4.962518 | 5.217674 |
| mesh | 5.428335 | 5.761123 |
| mesh.pretty | 6.939520 | 7.500490 |
| numbers | 6.307731 | 6.701964 |
| random | 4.366775 | 4.493627 |
| twitter | 5.991603 | 6.037428 |
| twitterescaped | 6.340564 | 6.449633 |
| update-center | 5.917625 | 6.072642 |

This is only on one type of machine (Kady Lake R). The least affected file is (interestingly) `twitter.json`, and even that got a close to 1% improvement (overall, that's basically a wash).