@jkeiser 2020-01-25
I see we have some performance degradations in checkperf on both arm and amd64. I still think it's worth thinking over: the test file we use is `twitter.json`, which is by *far* the least affected by this change. It gets < 1% where most files get 5-8%.