@lemire 2020-01-25
I am not sure that the failing tests are indicative of a genuine problem.