@lemire 2020-01-25
There are surprising high gains, but the gains are not all over. I am getting a perf. reduction in gsoc-2018 on x64 (on both AMD Rome and Intel Skylake). However, there are nice gains elsewhere. On ARM (small cores like Skylark), it seems to be a net negative, but the gap is small. This may have to do with the poorer branch predictions on these small ARM cores, but I am not sure. 

I think we need to assess this with care. It definitively looks appealing at a first glance.

I think we should take the time to draw a few tables and get a good collection of hard and reliable numbers on different systems.