@jkeiser 2020-01-25
Agreed. I have a little less certainty about this patch than others.

It actually makes sense to add a feature detector similar to the UTF-8 and structural feature detector. Then compare *that* across platforms. By adding essentially a fourth branch to stage 1 (the loop, utf-8, and structurals < 8), we are increasing branch chaos for sure.

My current experiment is to see what happens if we parse strings *before* doing anything else. Basically only handle backslashes and quotes once, pass in_string to the structural detector, and never special case strings again. In that scenario, it may actually be more of a win because we'll be doing more work on each backslash.