@jkeiser 2020-01-25
Got some better data here on the tradeoff we're making.

I added escape feature benchmarking to `benchfeatures`. "Base" indicates the cost of a 0-structural block, 7-struct is a block with 7 structurals and 3 strings, and escape is a block with 7 structurals, 3 strings, . These three costs are cumulative: each builds on each other so you only see the extra cost for the feature. The branch miss % (which I couldn't report for my Kady Lake R machine because windows) shows you the percentage of *actual* branch misses versus *expected*--so if that % is bigger, the cost per block is actually smaller, and vice versa.

| Architecture       | Branch       |     Base | 7 Struct |   Escape |   7 Struct Miss |     Escape Miss |
|--------------------|--------------|----------|----------|----------|-----------------|-----------------|
| AVX2 Kady Lake R   | master       |     6.26 |     2.52 |        0 |            1.94 |               0 |
| AVX2 Kady Lake R   | if_backslash |      5.5 |      2.6 |        1 |            1.36 |            2.16 |
| AVX2 Skylake       | master       |     6.96 |     2.91 |        0 |      1.91 (80%) |               0 |
| AVX2 Skylake       | if_backslash |     6.12 |     2.76 |     1.23 |      2.08 (90%) |      2.47 (29%) |
| AVX2 Skylake X     | master       |     9.06 |     3.85 |        0 |      1.22 (30%) |               0 |
| AVX2 Skylake X     | if_backslash |     7.87 |     3.82 |     1.57 |      0.94 (95%) |      ??? -1.58 (49%) and 4.83 (45%) |
| AMD64 Ampere       | master       |     64.5 |     8.46 |        0 |     4.01 (129%) |               0 |
| AMD64 Ampere       | if_backslash |     64.5 |     8.44 |     1.56 |     3.98 (130%) |      11.1 (127%) |

## Conclusions

* On Intel, this seems to be the tradeoff:
  - +10% throughput and increased miss rate for "normal" blocks
  - -15-30% throughput for backslashes, including an abnormally low miss rate (might mean real miss cost is 15-90% for backslashes)
* On ARM, there's something weird going on. We pay the same base cost no matter what, so the tradeoff is guaranteed to be negative. My only guess here is either compiler screwed up and wasted instructions, or we are so superscalar on Ampere (heh) that all 9 escape instructions are free.
* The miss rates are pretty darn consistent across runs, which is interesting.
* Skylake-X had a consistent miss rate but weirdly fluctuating miss cost for escapes. Everything else was consistent. I don't know what's up there yet.

## Methodology

```
git checkout jkeiser/benchfeatures-escape
make run_benchfeatures
git checkout jkeiser/if-backslash-bench
make run_benchfeatures
```

I ran `run_benchfeatures` multiple times to ensure the numbers came out consistent. On Skylake-X, they didn't, so I ran `./benchfeatures -n 4000 -i 200` instead and got consistent numbers out of that.