@jkeiser 2020-01-25
If those numbers hold up for the rest of reality, the conclusion is if you believe backslashes comprise 33% or less of real world input (i.e. backslashes happen every 192 bytes or more), you want to put the `if` statement in. In our files, here are the numbers (backslash flips = places where one block has a backslash and the next does not, or vice versa--think of it as "miss rate for a naive branch predictor that always guesses whatever came last"):

| File          | % blocks with backslash | % backslash flips |
|---------------|------|------|
| apache_builds | 0.5% | 0.1% |
| canada | 0.0% | 0.0% |
| citm_catalog | 0.0% | 0.0% |
| github_events | 6.4% | 5.7% |
| gsoc-2018 | 15.9% | 22.4% |
| instruments | 0.0% | 0.0% |
| marine_ik | 0.0% | 0.0% |
| mesh | 0.0% | 0.0% |
| mesh.pretty | 0.0% | 0.0% |
| numbers | 0.0% | 0.0% |
| random | 0.0% | 0.0% |
| twitter | 6.1% | 6.5% |
| twitterescaped | 43.5% | 10.2% |
| update-center | 2.9% | 5.3% |

Only twitterescaped comes close to that threshold. This does accord with my experience of JSON REST APIs: even when there are some fields that have lots of backslashes, there's usually a lot of *other* data without it.