@lemire 2020-01-27
I am relaunching the fuzzer. This change should not introduce bugs, but let us be sure.