@jkeiser 2020-06-12
@lemire I finally got around to updating this PR for the modern world. On skylake, it definitely improves things on all files except gsoc-2018. It **hugely** increases the number of branch misses, though. It's going to eat into our branch budget something fierce. I think we should check it in (once I deal with ARM), but we'll want to keep it in mind if we find better places we want to insert branches.

| File                                | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. | master Branch Misses | PR Branch Misses | -Branch Misses | master Cache Misses | PR Cache Misses | -Cache Misses | master Cache Refs | PR Cache Refs | -Cache Refs |
| ---                                 |   --: |   --: |   --: |   --: |   --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |
| citm_catalog                        |  36.9 |  34.0 |    7% | 139.2 | 130.8 |    6% |    144 |    174 |  -20% |    652 |    101 |   84% |  72062 |  71625 |    0% |
| mesh.pretty                         |  38.5 |  35.7 |    7% | 146.2 | 137.7 |    5% |     31 |     43 |  -38% |     83 |     54 |   34% |  69648 |  69359 |    0% |
| twitter_api_response                |  41.4 |  38.4 |    7% | 148.1 | 140.4 |    5% |      2 |      1 |   50% |      0 |      0 |    0% |     23 |     99 | -330% |
| tree-pretty                         |  39.9 |  37.2 |    6% | 149.1 | 140.6 |    5% |      3 |      4 |  -33% |      0 |      0 |    0% |    445 |    442 |    0% |
| github_events                       |  37.0 |  34.5 |    6% | 137.3 | 129.9 |    5% |      6 |      7 |  -16% |      0 |      0 |    0% |   2050 |   1344 |   34% |
| google_maps_api_response            |  39.3 |  36.7 |    6% | 145.7 | 137.2 |    5% |      2 |      3 |  -50% |      0 |      0 |    0% |     65 |    164 | -152% |
| apache_builds                       |  39.5 |  37.0 |    6% | 147.8 | 139.3 |    5% |      9 |     37 | -311% |      0 |      0 |    0% |   4449 |   4840 |   -8% |
| canada                              |  48.4 |  45.5 |    5% | 166.1 | 157.6 |    5% |   5165 |   5489 |   -6% |  13736 |  13026 |    5% | 115684 | 115705 |    0% |
| google_maps_api_compact_response    |  54.0 |  51.2 |    5% | 198.8 | 190.3 |    4% |      2 |      3 |  -50% |      0 |      0 |    0% |     76 |     79 |   -3% |
| twitter_api_compact_response        |  45.4 |  43.1 |    4% | 160.8 | 153.5 |    4% |      2 |      3 |  -50% |      0 |      0 |    0% |     20 |     61 | -204% |
| marine_ik                           |  57.2 |  54.4 |    4% | 189.4 | 180.9 |    4% |   8457 |   8616 |   -1% |  77692 |  78419 |    0% | 186584 | 186416 |    0% |
| repeat                              |  55.6 |  52.9 |    4% | 197.8 | 192.2 |    2% |      2 |      4 | -100% |     10 |      0 |    0% |    106 |      0 |    0% |
| numbers                             |  40.5 |  38.7 |    4% | 152.7 | 144.3 |    5% |     13 |    109 | -738% |      0 |      0 |    0% |   6496 |   7273 |  -11% |
| random                              |  60.7 |  58.1 |    4% | 219.1 | 212.9 |    2% |    772 |    989 |  -28% |      0 |      0 |    0% |  27809 |  27634 |    0% |
| mesh                                |  48.8 |  46.7 |    4% | 182.7 | 174.2 |    4% |    588 |    824 |  -40% |      0 |      9 |    0% |  43255 |  42911 |    0% |
| instruments                         |  40.1 |  38.6 |    3% | 151.6 | 143.1 |    5% |     17 |    197 | -1058% |      0 |      0 |    0% |  10578 |  10486 |    0% |
| twitter                             |  44.8 |  43.2 |    3% | 156.8 | 150.1 |    4% |   1135 |   1639 |  -44% |      0 |      0 |    0% |  27125 |  27135 |    0% |
| twitterescaped                      |  42.7 |  41.6 |    2% | 139.8 | 138.5 |    0% |   1897 |   1858 |    2% |      0 |      0 |    0% |  25159 |  24998 |    0% |
| twitter_timeline                    |  41.9 |  41.0 |    2% | 156.2 | 153.9 |    1% |      6 |      7 |  -16% |      0 |      0 |    0% |    852 |    450 |   47% |
| update-center                       |  45.4 |  45.2 |    0% | 150.2 | 142.2 |    5% |   1965 |   2861 |  -45% |      0 |      0 |    0% |  25236 |  25264 |    0% |
| gsoc-2018                           |  34.1 |  36.0 |   -5% | 107.1 | 101.3 |    5% |   5999 |  14874 | -147% |  24338 |  23854 |    1% | 115983 | 116436 |    0% |