@jkeiser 2020-06-16
@lemire yeah, that's probably the right thing. I double checked and we do get a real benefit by disabling this extra check on ARM. PR-1 is the old version with the backslash check, and PR is the current PR, which disables it on ARM. This is ampere, which is literally universally better off without the if check (it does run more instructions but we knew that):

| File                                | PR-1 Cycles | PR Cycles | +Throughput | PR-1 Instr. | PR Instr. | -Instr. | PR-1Branch Misses | PR Branch Misses | -Branch Misses | PR-1 Cache Misses | PR Cache Misses | -Cache Misses | master Cache Refs | PR Cache Refs | -Cache Refs |
| ---                                 |   --: |   --: |   --: |   --: |   --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |
| gsoc-2018                           | 234.4 | 225.2 |    3% | 141.1 | 145.0 |   -2% |  20673 |  10546 |   48% |  52037 |  52037 |    0% | 575183 | 571318 |    0% |
| twitter_timeline                    | 274.8 | 264.1 |    3% | 197.3 | 199.7 |   -1% |    510 |    375 |   26% |    693 |    696 |    0% |  13003 |  13004 |    0% |
| twitter_api_compact_response        | 282.2 | 275.6 |    2% | 200.9 | 205.5 |   -2% |     74 |     55 |   25% |    116 |    113 |    2% |   3247 |   3238 |    0% |
| twitterescaped                      | 256.2 | 251.0 |    2% | 179.3 | 181.2 |   -1% |   4355 |   4379 |    0% |   8828 |   8832 |    0% | 144785 | 145651 |    0% |
| update-center                       | 265.0 | 259.8 |    1% | 188.2 | 193.0 |   -2% |   4603 |   4125 |   10% |   8381 |   8387 |    0% | 153726 | 153789 |    0% |
| google_maps_api_compact_response    | 288.4 | 283.0 |    1% | 240.5 | 245.5 |   -2% |    140 |    130 |    7% |    178 |    186 |   -4% |   4948 |   4936 |    0% |
| github_events                       | 256.7 | 251.9 |    1% | 173.6 | 178.2 |   -2% |    291 |    267 |    8% |   1059 |   1058 |    0% |  16425 |  16473 |    0% |
| twitter_api_response                | 271.0 | 266.3 |    1% | 185.6 | 190.3 |   -2% |     76 |     69 |    9% |    238 |    233 |    2% |   4331 |   4334 |    0% |
| twitter                             | 303.2 | 298.0 |    1% | 197.9 | 202.5 |   -2% |   3755 |   3206 |   14% |   9916 |   9914 |    0% | 166835 | 166826 |    0% |
| numbers                             | 263.0 | 259.0 |    1% | 190.6 | 195.6 |   -2% |    635 |    591 |    6% |   2386 |   2384 |    0% |  44096 |  44085 |    0% |
| google_maps_api_response            | 261.0 | 257.2 |    1% | 182.4 | 187.4 |   -2% |     81 |     67 |   17% |    430 |    433 |    0% |   7271 |   7268 |    0% |
| marine_ik                           | 272.8 | 268.8 |    1% | 232.0 | 237.0 |   -2% |  13517 |  12813 |    5% |  46656 |  46655 |    0% | 1143932 | 1143917 |    0% |
| mesh.pretty                         | 255.1 | 251.4 |    1% | 183.0 | 188.0 |   -2% |    682 |     38 |   94% |  24685 |  24684 |    0% | 430724 | 430714 |    0% |
| instruments                         | 261.7 | 257.9 |    1% | 189.3 | 194.3 |   -2% |    792 |    767 |    3% |   3484 |   3482 |    0% |  63893 |  63893 |    0% |
| mesh                                | 270.9 | 267.0 |    1% | 223.1 | 228.1 |   -2% |   1061 |   1024 |    3% |  11344 |  11345 |    0% | 254416 | 254419 |    0% |
| citm_catalog                        | 254.4 | 251.0 |    1% | 174.9 | 179.9 |   -2% |    459 |    466 |   -1% |  27031 |  27031 |    0% | 434832 | 434794 |    0% |
| apache_builds                       | 260.9 | 257.4 |    1% | 184.8 | 189.8 |   -2% |    409 |    433 |   -5% |   2027 |   2028 |    0% |  35506 |  35514 |    0% |
| tree-pretty                         | 261.9 | 258.7 |    1% | 186.3 | 191.3 |   -2% |    125 |    116 |    7% |    578 |    575 |    0% |   9905 |   9905 |    0% |
| random                              | 394.3 | 389.7 |    1% | 272.5 | 277.5 |   -1% |   4082 |   3715 |    8% |   8023 |   8023 |    0% | 189962 | 189972 |    0% |
| canada                              | 263.9 | 261.1 |    1% | 206.2 | 211.2 |   -2% |   9839 |  11002 |  -11% |  35214 |  35212 |    0% | 748109 | 748082 |    0% |
| repeat                              | 414.1 | 411.2 |    0% | 248.8 | 253.8 |   -2% |     40 |     36 |   10% |    135 |    135 |    0% |   3376 |   3375 |    0% |